poiu

+++ application/controllers/Users.php
@@ -1,9 +1,9 @@
 <?php
 defined('BASEPATH') OR exit('No direct script access allowed');
 
-require_once 'My_Controller.php';
 
-class Users extends My_Controller {
+
+class Users extends MY_Controller {
 
     public function __construct() {
         parent::__construct();
@@ -67,7 +67,7 @@ class Users extends My_Controller {
             $data = [
                 'title' => 'Adicionar Endereço',
                 'styles' => ['style'],
-                'scripts' => ['ajxCadastra', 'mascaras']
+                'scripts' => ['endereco']
             ];
             
             $this->my_header($data);
@@ -151,7 +151,7 @@ class Users extends My_Controller {
         } else {
             $this->my_header([
                 'title' => "Cadastro", 
-                'scripts' => ['ajxCadastra','mascaras'], 
+                'scripts' => ['cadastro'], 
                 'styles' => ['style']
             ]);
             $this->load->view('cadastro');
@@ -164,22 +164,7 @@ class Users extends My_Controller {
             redirect('entrar');
         } 
         if($this->input->server('REQUEST_METHOD') == 'POST') {
-            $id = $this->uri->segment(2);
-            //Apagando registro diretamente
-                // 
-                // $del = $this->db->delete('users', [ 'id' => $id]);
-    
-                // if ($del) {
-                //     echo json_encode([
-                //         'msg' => "Registro apagado com sucesso!"
-                //     ]);
-                // } else { 
-                //     echo json_encode([
-                //         'msg' => "O registro não foi apagado!"
-                //     ]);
-                // }
-            
-            // Apagando com coluna st_usuario (0 ou 1)    
+            $id = $this->uri->segment(2);  
             $del = $this->db->set('st_usuario', 0)
                             ->where('id', $id)
                             ->update('users');      
@@ -199,90 +184,16 @@ class Users extends My_Controller {
 
     public function editUser() {
         if (!$this->session->userdata('user_id')) {
-            redirect('entrar');
+            redirect(base_url('entrar'));
         } 
+        if (!($this->session->userdata('user_tipo_id') == 1)){ 
+            redirect(base_url(''));
+        }
         $id = $this->uri->segment(2);
         if($this->input->server('REQUEST_METHOD') == 'POST') {
             $attData = [];
-            $this->form_validation->set_rules('user' , 'Usuário', 'required');
-            $this->form_validation->set_rules('email' , 'Email', 'required|valid_email', [
-                'is_unique' => "E-mail já cadastrado."
-            ]);
-            $this->form_validation->set_rules('doc-cpf-cnpj' , 'Documento', 'required');
-
-            if($this->input->post('senha')) { 
-                $this->form_validation->set_rules('senha' , 'Senha', 'required|min_length[6]');
-                $this->form_validation->set_rules('confirmacao-senha' , 'Repita a senha', 'required|matches[senha]');
-                $attData['senha'] = $this->input->post('senha'); 
-            }
-
-            if ($this->input->post('tipo-documento') == 1) { 
-                $this->form_validation->set_rules('doc-cpf-cnpj', 'Documento', 'required|min_length[14]', [
-                    'min_length' => "Formato do CPF não é válido!"
-                ]);
-            }else if($this->input->post('tipo-documento') == 2) {
-                $this->form_validation->set_rules('doc-cpf-cnpj', 'Documento', 'required|min_length[18]', [
-                    'min_length' => "Formato do CNPJ não é válido!"
-                ]);
-            } else {
-                $this->form_validation->set_rules('doc-cpf-cnpj' , 'Documento', 'required');
-            }
-
-            if($this->form_validation->run() == FALSE) {
-                $this->output->set_status_header(400);
-                $this->form_validation->set_error_delimiters('','');
-                echo json_encode([
-                    'error_nome' => form_error('nome'),
-                    'error_email' => form_error('email'),
-                    'error_senha' => form_error('senha'),
-                    'error_cs' => form_error('confirmacao-senha'),
-                    'error_doc' => form_error('doc-cpf-cnpj'),
-                    'csrf' => $this->security->get_csrf_hash()
-                ]);
-            } else { 
-                $attData = array_merge($attData, [
-                    'nome' => $this->input->post('user'),
-                    'email' => $this->input->post('email'),
-                    'tipo_documento' => $this->input->post('tipo-documento'),
-                    'doc_cpf_cnpj' => preg_replace("/[^0-9]/", "", $this->input->post('doc-cpf-cnpj'))
-                ]);
-                $att = $this->db->set($attData)
-                ->where('id', $id)
-                ->update('users');      
-                if($att) { 
-                    echo json_encode([
-                        'msg' => "Registro editado com sucesso!",
-                    ]);
-                } else { 
-                    echo json_encode([
-                        'msg' => "O registro não foi editado!",
-                        'csrf' => $this->security->get_csrf_hash()
-                    ]);
-                }
-            }
-        }else { 
-            $data['edit_user'] = $this->db->select("id , nome, email, img_profile_path as caminho_foto, doc_cpf_cnpj as doc,tipo_documento")
-                                        ->from('users')
-                                        ->where('id', $id)
-                                        ->get()->result();
-            $data = array_merge($data, [
-                'title' => 'Editar Usuário', 
-                'scripts' => ['ajxEdita','mascaras'],
-                'styles' => ['style']
-            ]);
-            $this->my_header($data);
-            $this->load->view('editar');
-            $this->load->view('footer/footer');
-        }
-    }
-
-    public function editprofile() { 
-        if (!$this->session->userdata('user_id')) {
-            redirect('entrar');
-        } 
-        if($this->input->server('REQUEST_METHOD') == 'POST') {
-            $attData = [];
             $errors = [];
+            $imgDelete = FALSE;
 
             $config['upload_path']          = "assets/imgs/";
             $config['allowed_types']        = 'gif|jpg|png|jpeg';
@@ -297,7 +208,7 @@ class Users extends My_Controller {
             if($this->input->post('senha')) { 
                 $this->form_validation->set_rules('senha' , 'Senha', 'required|min_length[6]');
                 $this->form_validation->set_rules('confirmacao-senha' , 'Repita a senha', 'required|matches[senha]');
-                $attData['senha'] = $this->input->post('senha'); 
+                $attData['senha'] = password_hash($this->input->post('senha'), PASSWORD_DEFAULT); 
             }
             
             if($_FILES['foto']['name']) {
@@ -305,8 +216,21 @@ class Users extends My_Controller {
                     $errors['error_foto'] = $this->upload->display_errors('','');
                 } else { 
                     $attData['img_profile_path'] = $config['upload_path'] . $this->upload->data('file_name');
+                    $imgDelete = TRUE;
                 } 
             } 
+
+            if ($this->input->post('tipo-documento') == 1) { 
+                $this->form_validation->set_rules('doc-cpf-cnpj', 'Documento', 'required|min_length[14]', [
+                    'min_length' => "Formato do CPF não é válido!"
+                ]);
+            }else if($this->input->post('tipo-documento') == 2) {
+                $this->form_validation->set_rules('doc-cpf-cnpj', 'Documento', 'required|min_length[18]', [
+                    'min_length' => "Formato do CNPJ não é válido!"
+                ]);
+            } else {
+                $this->form_validation->set_rules('doc-cpf-cnpj' , 'Documento', 'required');
+            }
 
             if($this->form_validation->run() == FALSE || isset($errors['error_foto'])) {
                 $this->output->set_status_header(400);
@@ -316,23 +240,28 @@ class Users extends My_Controller {
                     'error_email' => form_error('email'),
                     'error_senha' => form_error('senha'),
                     'error_cs' => form_error('confirmacao-senha'),
+                    'error_doc' => form_error('doc-cpf-cnpj'),
                     'csrf' => $this->security->get_csrf_hash()
                 ]);
                 echo json_encode($errors);
             } else { 
-                $query = $this->db->select('img_profile_path as caminho_antigo')->from('users')->where('id', $this->session->userdata('user_id'))->get()->result();
+                $query = $this->db->select('img_profile_path as caminho_antigo')->from('users')->where('id', $id)->get()->result();
                 list($old) = $query;
-                unlink($old->caminho_antigo);
+                if($old->caminho_antigo && $imgDelete){ 
+                    unlink($old->caminho_antigo);
+                }
                 $attData = array_merge($attData, [
                     'nome' => $this->input->post('user'),
-                    'email' => $this->input->post('email') 
+                    'email' => $this->input->post('email') ,
+                    'tipo_documento' => $this->input->post('tipo-documento'),
+                    'doc_cpf_cnpj' => preg_replace("/[^0-9]/", "", $this->input->post('doc-cpf-cnpj'))
                 ]);
                 $att = $this->db->set($attData)
-                ->where('id', $this->session->userdata('user_id'))
+                ->where('id', $id)
                 ->update('users');      
                 if($att) { 
                     echo json_encode([
-                        'msg' => "Registro editado com sucesso!",
+                        'msg' => "Registro editado com sucesso!"
                     ]);
                 } else { 
                     echo json_encode([
@@ -342,20 +271,119 @@ class Users extends My_Controller {
                 }
             }
         }else { 
-            $data['profile_data'] = $this->db->select("nome, email, img_profile_path as caminho_foto ")
+            $data['profile_data'] = $this->db->select("id, nome, email, img_profile_path as caminho_foto, doc_cpf_cnpj as doc,tipo_documento")
                                         ->from('users')
-                                        ->where('id', $this->session->userdata('user_id'))
+                                        ->where('id', $id)
                                         ->get()->result();
             $data = array_merge($data, [
                 'title' => 'Editar Perfil',
                 'styles' => ['style'],
-                'scripts' => ['ajxEdita']
+                'scripts' => ['editar']
+            ]);
+            $this->my_header($data);
+            $this->load->view('editar');
+            $this->load->view('footer/footer');
+        }
+    }
+
+    public function editprofile() { 
+        if (!$this->session->userdata('user_id')) {
+            redirect(base_url('entrar'));
+        } 
+        $id = $this->session->userdata('user_id');
+        if($this->input->server('REQUEST_METHOD') == 'POST') {
+            $attData = [];
+            $errors = [];
+            $imgDelete = FALSE;
+
+            $config['upload_path']          = "assets/imgs/";
+            $config['allowed_types']        = 'gif|jpg|png|jpeg';
+
+            $this->load->library('upload', $config);
+
+
+            $this->form_validation->set_rules('user' , 'Usuário', 'required');
+            $this->form_validation->set_rules('email' , 'Email', 'required|valid_email', [
+                'is_unique' => "E-mail já cadastrado."
+            ]);
+            if($this->input->post('senha')) { 
+                $this->form_validation->set_rules('senha' , 'Senha', 'required|min_length[6]');
+                $this->form_validation->set_rules('confirmacao-senha' , 'Repita a senha', 'required|matches[senha]');
+                $attData['senha'] = password_hash($this->input->post('senha'), PASSWORD_DEFAULT); 
+            }
+            
+            if($_FILES['foto']['name']) {
+                if(!$this->upload->do_upload('foto')) {
+                    $errors['error_foto'] = $this->upload->display_errors('','');
+                } else { 
+                    $attData['img_profile_path'] = $config['upload_path'] . $this->upload->data('file_name');
+                    $imgDelete = TRUE;
+                } 
+            } 
+
+            if ($this->input->post('tipo-documento') == 1) { 
+                $this->form_validation->set_rules('doc-cpf-cnpj', 'Documento', 'required|min_length[14]', [
+                    'min_length' => "Formato do CPF não é válido!"
+                ]);
+            }else if($this->input->post('tipo-documento') == 2) {
+                $this->form_validation->set_rules('doc-cpf-cnpj', 'Documento', 'required|min_length[18]', [
+                    'min_length' => "Formato do CNPJ não é válido!"
+                ]);
+            } else {
+                $this->form_validation->set_rules('doc-cpf-cnpj' , 'Documento', 'required');
+            }
+
+            if($this->form_validation->run() == FALSE || isset($errors['error_foto'])) {
+                $this->output->set_status_header(400);
+                $this->form_validation->set_error_delimiters('','');
+                $errors = array_merge($errors, [
+                    'error_nome' => form_error('nome'),
+                    'error_email' => form_error('email'),
+                    'error_senha' => form_error('senha'),
+                    'error_cs' => form_error('confirmacao-senha'),
+                    'error_doc' => form_error('doc-cpf-cnpj'),
+                    'csrf' => $this->security->get_csrf_hash()
+                ]);
+                echo json_encode($errors);
+            } else { 
+                $query = $this->db->select('img_profile_path as caminho_antigo')->from('users')->where('id', $id)->get()->result();
+                list($old) = $query;
+                if($old->caminho_antigo && $imgDelete){ 
+                    unlink($old->caminho_antigo);
+                }
+                $attData = array_merge($attData, [
+                    'nome' => $this->input->post('user'),
+                    'email' => $this->input->post('email') ,
+                    'tipo_documento' => $this->input->post('tipo-documento'),
+                    'doc_cpf_cnpj' => preg_replace("/[^0-9]/", "", $this->input->post('doc-cpf-cnpj'))
+                ]);
+                $att = $this->db->set($attData)
+                ->where('id', $id)
+                ->update('users');      
+                if($att) { 
+                    echo json_encode([
+                        'msg' => "Registro editado com sucesso!"
+                    ]);
+                } else { 
+                    echo json_encode([
+                        'msg' => "O registro não foi editado!",
+                        'csrf' => $this->security->get_csrf_hash()
+                    ]);
+                }
+            }
+        }else { 
+            $data['profile_data'] = $this->db->select("id, nome, email, img_profile_path as caminho_foto, doc_cpf_cnpj as doc,tipo_documento")
+                                        ->from('users')
+                                        ->where('id', $id)
+                                        ->get()->result();
+            $data = array_merge($data, [
+                'title' => 'Editar Perfil',
+                'styles' => ['style'],
+                'scripts' => ['editarperfil']
             ]);
             $this->my_header($data);
             $this->load->view('editarperfil');
             $this->load->view('footer/footer');
         }
-
-        
     }
 }
